+++ src/main/java/Command/MusicModule/RadioCommand.java
@@ -12,6 +12,7 @@ import Command.Command;
 import static Command.Command.embed;
 import Constants.Constants;
 import Setting.Prefix;
+import Utility.UtilBot;
 import Utility.UtilNum;
 import java.awt.Color;
 import java.time.Instant;
@@ -58,7 +59,7 @@ public class RadioCommand implements Command{
             embedpl.addField("Available Radio Stations", stations, true);
             embedpl.setThumbnail(Constants.B_AVATAR);
             embedpl.setFooter("Requested by " + e.getAuthor().getName(), e.getAuthor().getEffectiveAvatarUrl());
-            embedpl.setColor(UtilNum.randomColor());
+            embedpl.setColor(UtilBot.randomColor());
             embedpl.setTimestamp(Instant.now());
             
             e.getChannel().sendMessage(embedpl.build()).queue();